requirements.txt
amqp==5.1.1
annotated-types==0.6.0
ansi2html==1.8.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
cachelib==0.9.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloveriiif==0.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
dash==2.13.0
dash-ag-grid==2.4.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
distinctipy==1.2.3
dnspython==2.4.2
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
fiona==1.9.5
Flask==2.2.5
Flask-Caching==2.0.2
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.10.0
GeoAlchemy==0.7.2
geopandas==0.14.1
girder-client==3.1.24
gunicorn==21.2.0
idna==3.4
imagecodecs==2023.9.18
imageio==2.32.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multiprocess==0.70.15
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook_shim==0.2.3
numpy==1.26.0
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pgvector==0.2.4
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
shapely==2.0.2
SimpleITK==2.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
sqlmodel==0.0.14
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.221
uri-template==1.3.0
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.3
